# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management API",
        "description": "API for managing school operations including students, teachers, courses, grades, and attendance",
        "contact": {
            "email": "admin@school.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
        "/api/asignaturas": {
            "get": {
                "tags": [
                    "Asignaturas"
                ],
                "summary": "Obtiene la lista de Asignaturas.",
                "operationId": "2d427742471a5cd10d43d7965334303d",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Filtrar por Nombre de asignatura.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "docente",
                        "in": "query",
                        "description": "Filtrar por Docente asignado a la asignatura.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de asignaturas obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre_asignatura": {
                                                "type": "string",
                                                "example": "Matemáticas"
                                            },
                                            "docente": {
                                                "type": "string",
                                                "example": "Carlos Garcia"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Asignaturas"
                ],
                "summary": "Crea una nueva asignatura.",
                "operationId": "f5befe794e764a502a9059dfec129ae5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "credits",
                                    "teacher_id",
                                    "grade_level"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Matemáticas"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "MAT101"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Curso introductorio a las matemáticas."
                                    },
                                    "credits": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "teacher_id": {
                                        "description": "ID del docente que imparte la asignatura.",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "grade_level": {
                                        "type": "string",
                                        "example": "10th Grade"
                                    },
                                    "schedule": {
                                        "type": "string",
                                        "example": "Lunes, Miércoles 10:00-11:30"
                                    },
                                    "room": {
                                        "type": "string",
                                        "example": "Aula 205"
                                    },
                                    "max_students": {
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Asignatura creada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Asignatura created successfully"
                                        },
                                        "asignatura": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre_asignatura": {
                                                    "type": "string",
                                                    "example": "Matemáticas"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/asignaturas/{id}": {
            "put": {
                "tags": [
                    "Asignaturas"
                ],
                "summary": "Permite editar una asignatura.",
                "operationId": "96cff0fee053aaf9afda783d217f1536",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la asignatura a actualizar.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Matemáticas Avanzadas"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "MAT101"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Curso avanzado de matemáticas."
                                    },
                                    "credits": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "teacher_id": {
                                        "description": "ID del docente que imparte la asignatura.",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "grade_level": {
                                        "type": "string",
                                        "example": "11th Grade"
                                    },
                                    "schedule": {
                                        "type": "string",
                                        "example": "Martes, Jueves 09:00-10:30"
                                    },
                                    "room": {
                                        "type": "string",
                                        "example": "Aula 301"
                                    },
                                    "max_students": {
                                        "type": "integer",
                                        "example": 25
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Asignatura actualizada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Asignatura updated successfully"
                                        },
                                        "asignatura": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre_asignatura": {
                                                    "type": "string",
                                                    "example": "Matemáticas Avanzadas"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asignatura no encontrada."
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}